+++ lib/ui/widgets/library/folder_detail_view.dart
@@ -55,6 +55,7 @@ class _FolderDetailViewState extends ConsumerState<FolderDetailView> {
       }
     });
     WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (!mounted) return;
       _enrichMissingArtwork();
     });
   }